requirements.txt
langchain
langchain-google-genai
streamlit
pydantic
python-dotenv
google-generativeai
pytesseract
pillow
pyaudio
gtts
playsound
SpeechRecognition
pydub
googletrans


